pyproject.toml
[project]
name = "tech-teach"
version = "0.1.0"
description = "AI-Powered Teaching Assistant using Google ADK Agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "google-genai>=1.43.0",
    "google-adk>=0.1.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.37.0",
    "tenacity>=9.1.2",
    "reportlab>=4.0.0",
    "pillow>=10.0.0",
    "aiofiles>=23.0.0",
    "pypdf>=4.0.0",
    "imageio>=2.30.0",
    "imageio-ffmpeg>=0.4.0",
    "numpy>=1.24.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
